float support added

+++ SimpleCV/ImageClass.py
@@ -4556,7 +4556,7 @@ class Image:
         >>> newimg.show()
         """
         if type(other) != type(self):
-            newnpimg = self.getNumpy()/other
+            newnpimg = (self.getNumpy()/other).astype(np.uint8)
             return Image(newnpimg, colorSpace=self._colorSpace)
         if self.size() != other.size():
             warnings.warn("Both images should have same dimensions. Returning None.")
@@ -4592,7 +4592,7 @@ class Image:
         >>> newimg.show()
         """
         if type(other) != type(self):
-            newnpimg = other*self.getNumpy()
+            newnpimg = (other*self.getNumpy()).astype(np.uint8)
             return Image(newnpimg, colorSpace=self._colorSpace)
         if self.size() != other.size():
             warnings.warn("Both images should have same dimensions. Returning None.")
